requirements.txt
streamlit==1.36.0
langchain==0.2.10
langchain-anthropic==0.1.20
langchain-community==0.2.9
langchain-core==0.2.22
langchain-google-community==1.0.7
langchain-google-genai==1.0.8
langchain-google-vertexai==1.0.7
langchain-openai==0.1.17
pydantic==2.8.2
bandit==1.7.9
pycodestyle==2.12.0
GitPython==3.1.43
faiss-cpu==1.8.0.post1
